refactor: move database to storage/internal package
