+++ pkg/storage/internal/database/database_test.go
@@ -1,12 +1,12 @@
 package database_test
 
 import (
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/storage/internal/database"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/storage/internal/database/models"
 	"testing"
 	"time"
 
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/defs"
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/database"
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/database/models"
 	"github.com/stretchr/testify/require"
 )
